feat: make main window larger on 1440p screens
When browser screeen in atleast 2000px in width the main window along
with the image will scale up a bit.

Because I want the image to be pre-scaled for the resolution it's going
to be displayed at I needed to provide 2 images.

Using the <picture> tag along with a <source> with a min-width solves
the issue of what image is displayed.

because webp is nice i went and changed the images to webp.
the highest-res image I have is also provided now.